% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_plotDataTab.R
\name{barTable}
\alias{barTable}
\title{Generate the corresponding table of the Barchart}
\usage{
barTable(data, pvalue, fchange)
}
\arguments{
\item{data}{Differential Expression results (deTable)}

\item{pvalue}{P-value threshold}

\item{fchange}{Fold-Change threshold}
}
\value{
Returns the table to be used in the Barchart
}
\description{
Generate the corresponding table of the Barchart
}
