% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_filterDataTab.R
\name{batchComBat}
\alias{batchComBat}
\title{ComBat Batch Effect Correction}
\usage{
batchComBat(readCounts, meta)
}
\arguments{
\item{meta, }{Metadata Table}

\item{counts, }{Filtered Count Table}
}
\value{
Returns batch-effect corrected table
}
\description{
ComBat Batch Effect Correction
}
