% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_filterDataTab.R
\name{batchHarman}
\alias{batchHarman}
\title{Harman Batch Effect Correction}
\usage{
batchHarman(counts, meta)
}
\arguments{
\item{counts, }{Filtered Count Table}

\item{meta, }{Metadata Table}
}
\value{
Returns batch-effect corrected table
}
\description{
Harman Batch Effect Correction
}
