% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_compDataTab.R
\name{bulk_compData}
\alias{bulk_compData}
\title{Bulk Compare Data Server}
\usage{
bulk_compData(input, output, session, rv, de)
}
\arguments{
\item{rv}{filtered counts}

\item{de}{DE results and replicates/conditions info}
}
\value{
None
}
\description{
Bulk Compare Data Server
}
