% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_footprintTab.R
\name{bulk_faData}
\alias{bulk_faData}
\title{Bulk FootPrint Tab Server}
\usage{
bulk_faData(input, output, session, counts, de)
}
\arguments{
\item{counts}{Unfiltered Count Table (Reactive Value)}

\item{de}{Differential Expression Results (Reactive Value)}
}
\value{
None
}
\description{
Bulk FootPrint Tab Server
}
