% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_filterDataTab.R
\name{bulk_filterData}
\alias{bulk_filterData}
\title{Bulk Filter Tab Server}
\usage{
bulk_filterData(input, output, session, counts)
}
\arguments{
\item{counts}{The loaded Count Table}
}
\value{
filt A reactive value with the filtered Count Table
}
\description{
Bulk Filter Tab Server
}
