% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_goDataTab.R
\name{bulk_goData}
\alias{bulk_goData}
\title{Bulk Functional Annotation Tab Server}
\usage{
bulk_goData(input, output, session, counts, de)
}
\arguments{
\item{counts}{Unfiltered Count Table (Reactive Value)}

\item{de}{Differential Expression Results (Reactive Value)}
}
\value{
None
}
\description{
Bulk Functional Annotation Tab Server
}
