% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_loadDataTab.R
\name{bulk_loadData}
\alias{bulk_loadData}
\title{Bulk load Data SERVER}
\usage{
bulk_loadData(input, output, session)
}
\value{
counts the loaded Count Table
}
\description{
Bulk load Data SERVER
}
