% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_plotDataTab.R
\name{bulk_plotData}
\alias{bulk_plotData}
\title{Bulk Plot Data Tab Server}
\usage{
bulk_plotData(input, output, session, rv)
}
\arguments{
\item{rv}{Reactive value containing the DE results (deTable)}
}
\value{
None
}
\description{
Bulk Plot Data Tab Server
}
