% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_footprintTab.R
\name{convert_gene_type}
\alias{convert_gene_type}
\title{Gene type converter function}
\usage{
convert_gene_type(de_data, input_type, output_type)
}
\arguments{
\item{de_data}{Table with de results}

\item{input_type}{input gene type e.g. ENSEMBL}

\item{output_type}{output gene type e.g. Symbol}
}
\value{
Returns a df with gene names convert to symbols
}
\description{
Gene type converter function
}
