% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_deDataTab.R
\name{deALDE}
\alias{deALDE}
\title{ALDEx2 Differential Expression Pipeline}
\usage{
deALDE(readCounts, conditionNo, replicateNo)
}
\arguments{
\item{readCounts}{The filtered CountTable}

\item{conditionNo}{The number of Conditions}

\item{replicateNo}{The number of Replicates}
}
\value{
ALDEx2 DE pipeline results (with normalized counts and logFC from Edge)
}
\description{
ALDEx2 Differential Expression Pipeline
}
