% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_deDataTab.R
\name{dePipelineCaller}
\alias{dePipelineCaller}
\title{Differential Expression Pipeline Caller}
\usage{
dePipelineCaller(filteredCounts, metaData, testType, paramX,
  selectedPackage, session, useBatch)
}
\arguments{
\item{testType}{The type of test to be used}

\item{paramX}{Fit type for DESeq2, and normalization method for limma and edgeR}

\item{selectedPackage}{The package(pipeline) to be executed}

\item{session}{Current R session}

\item{readCounts}{The filtered CountTable}
}
\value{
Returns the results of the appropriate DE pipeline
}
\description{
Function that calls the appropriate DGE pipeline
and provides exception handling
}
