% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_deDataTab.R
\name{deSequence}
\alias{deSequence}
\title{DESeq2 Differential Expression Pipeline}
\usage{
deSequence(readCounts, meta, testType, fitType, useBatch)
}
\arguments{
\item{readCounts}{The filtered CountTable}

\item{meta}{Metadata table}

\item{testType}{The type of test to be used}

\item{fitType}{The type of fitType to be used}

\item{useBatch}{If FALSE reasigns IDs to row.names}
}
\value{
Returns DESeq2 DE pipeline results
}
\description{
DESeq2 Differential Expression Pipeline
}
