% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_footprintTab.R
\name{fetch_regulons}
\alias{fetch_regulons}
\title{Fetch Dorothea Regulon}
\usage{
fetch_regulons(org, conf_list)
}
\arguments{
\item{org}{Human or Mouse}

\item{conf_list}{list_with_conf}
}
\value{
Returns Dorothea regulons
}
\description{
Fetch Dorothea Regulon
}
