% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_footprintTab.R
\name{fetch_tf_activities}
\alias{fetch_tf_activities}
\title{Transcription Factor activity with DoRothEA}
\usage{
fetch_tf_activities(de_data, reg_size, regulons, session)
}
\arguments{
\item{de_data}{Table with de results and symbols for gene type}

\item{reg_size}{minimun size of regulons}

\item{regulons}{Dorothea regulons}

\item{session}{currently running r session}
}
\value{
Returns a data frame with TF activities estimated from TF footprints
}
\description{
Transcription Factor activity with DoRothEA
}
