% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_filterDataTab.R
\name{filterFunction}
\alias{filterFunction}
\title{Filter Function}
\usage{
filterFunction(countTable, method, value)
}
\arguments{
\item{countTable}{The loaded Count Table}
}
\value{
countTable The filtered Count Table
}
\description{
Filter Function
}
