% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_deTab.R
\name{genePlot}
\alias{genePlot}
\title{Gene Plots across clusters}
\usage{
genePlot(finalData, plotType, geneName, session)
}
\arguments{
\item{finalData}{Seurat object with cluster data}

\item{plotType}{The desired plot type}

\item{geneName}{The name/symbol of the gene of interest}
}
\value{
Returns DE Gene Plots
}
\description{
Function with exception handling that enables comparison of genes
across the different clusters
}
