% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_compDataTab.R
\name{generateIntersect}
\alias{generateIntersect}
\title{Extract Intersects}
\usage{
generateIntersect(deList)
}
\arguments{
\item{deList}{list containing the DE Results of the 3 packages}
}
\value{
Returns a list with DE genes according to the different packages
}
\description{
Extract Intersects
}
