% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_loadDataTab.R
\name{generateSummary}
\alias{generateSummary}
\title{Generate Summary}
\usage{
generateSummary(counts, session)
}
\value{
df a dataframe containing the summary
}
\description{
Generates a summary of the columns(samples) in the Count Table
}
