% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_footprintTab.R
\name{get_pathway_activity}
\alias{get_pathway_activity}
\title{Fetch Pathway Activity}
\usage{
get_pathway_activity(de_data, organism, top, session)
}
\arguments{
\item{de_data}{de results dataframe}

\item{organism}{Human or Mouse}

\item{top}{num of top n genes for generating the model matrix sorted by p}

\item{session}{current R session}
}
\value{
a list with pathway activities, barplot, de_results matrix
}
\description{
Fetch Pathway Activity
}
