% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_footprintTab.R
\name{get_pathway_activity_per_sample}
\alias{get_pathway_activity_per_sample}
\title{Fetch Pathway Activity per Sample}
\usage{
get_pathway_activity_per_sample(de_data, organism, top, session)
}
\arguments{
\item{de_data}{de results dataframe}

\item{organism}{Human or Mouse}

\item{top}{# the top n genes for generating the model matrix sorted by p}
}
\value{
a list with pathway activity counts and a heatmap
}
\description{
Fetch Pathway Activity per Sample
}
