% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_compDataTab.R
\name{getAllDE}
\alias{getAllDE}
\title{Bulk Generate Data required to compare DE Package Results}
\usage{
getAllDE(readCounts, meta, batched)
}
\arguments{
\item{readCounts}{Filtered Counts Table}

\item{meta}{Metadata table}

\item{batched, }{whether batch effect was applied to the tables}
}
\value{
Returns a list with DE genes according to the different packages
}
\description{
Bulk Generate Data required to compare DE Package Results
}
