% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_deTab.R
\name{getClusterHeatmap}
\alias{getClusterHeatmap}
\title{Cluster Heatmap}
\usage{
getClusterHeatmap(s_object, markers, geneNo)
}
\arguments{
\item{s_object}{Seurat object with clustered data}

\item{markers}{Differential expression data}

\item{geneNo}{Number of genes to be displayed}
}
\value{
Returns DE Heatmap
}
\description{
Heatmap generated with Suerat
}
