% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_goDataTab.R
\name{getDEgenes}
\alias{getDEgenes}
\title{Get Differentially Expressed Genes (Bulk)}
\usage{
getDEgenes(data, type, pvalue, fchange, condition)
}
\arguments{
\item{data}{Differential Expression Reslts Table}

\item{type}{Filters by abs. significant, upregulation or downregulation}

\item{pvalue}{P-value threshold}

\item{fchange}{Fold-Change threshold}

\item{condition}{Filters by a given condition (AvB, BvC, etc.)}
}
\value{
Returns a vector with DE gene names
}
\description{
Filters the DE results table and returns the names of the DE genes
}
