% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_goDataTab.R
\name{getGenesetsGO}
\alias{getGenesetsGO}
\title{Get GO gene sets for GOSeq}
\usage{
getGenesetsGO(assayed.genes, genome, geneset_symbol)
}
\arguments{
\item{assayed.genes}{Genes in the DE set}

\item{genome}{Model organism}

\item{geneset_symbol}{Gene nomenclature}
}
\value{
Returns a data frame with two columns: genes and GO
}
\description{
Filters the DE results table and returns the names of the DE genes
}
