% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_plotDataTab.R
\name{getHeatData}
\alias{getHeatData}
\title{Returns Heatmap data}
\usage{
getHeatData(data, offset, replicates, geneNo, type, pvalue, fchange,
  conditionFC)
}
\arguments{
\item{data}{Differential Expression results (deTable)}

\item{offset}{The last column of DE results}

\item{replicates}{The number of replicates}

\item{geneNo}{The number of genes to be displayed}

\item{type}{Filter by up-, down-, or absolute significe DE}

\item{pvalue}{P-value threshold}

\item{fchange}{Fold-Change threshold}

\item{conditionFC}{The Fold-change column of the condition of interest}

\item{names}{Boolean - whether to show names or not}
}
\value{
table The data that corresponds to the displayed heatmap
}
\description{
Returns Heatmap data
}
