% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_compDataTab.R
\name{getIntersect}
\alias{getIntersect}
\title{Returns the Genes of a given intersect within the Venn Diagram}
\usage{
getIntersect(xlist, intersectID)
}
\arguments{
\item{xlist}{A list with DE genes according to the different packages}

\item{intersectID}{The corresponding number of the intersect of interest}
}
\value{
Returns the names of the genes within that intersect
}
\description{
Returns the Genes of a given intersect within the Venn Diagram
}
