% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_goDataTab.R
\name{goInfo}
\alias{goInfo}
\title{Generate GO Term Info}
\usage{
goInfo(goID, session)
}
\arguments{
\item{goID}{ID of the GO term of interest}

\item{session}{Current R session}
}
\value{
The concatinated GO TERM information
}
\description{
Calls GOTERM function and concatinates output
}
