% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_goDataTab.R
\name{histGoTerms}
\alias{histGoTerms}
\title{Generate GO Histogram}
\usage{
histGoTerms(data, gof, session)
}
\arguments{
\item{data}{Functional Annotation results}

\item{gof}{Boolean to use GO:Terms or GO:ID}
}
\value{
Returns a top 10 GO terms histogram
}
\description{
Creates a GO Histogram with GO terms or GO IDs
}
