% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_loadDataTab.R
\name{load10xData}
\alias{load10xData}
\title{Load 10x Data with TryCatch}
\usage{
load10xData(path, session)
}
\arguments{
\item{path}{The directory containing the 10X genomics data}
}
\value{
Returns 10x Genomics Data
}
\description{
Simply adds TryCatch statement to the function supplied by Seurat
}
