% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_normTab.R
\name{normalizeSeurat}
\alias{normalizeSeurat}
\title{Single Cell Normalize function}
\usage{
normalizeSeurat(s_object, normalizeMet, scaleF, varianceMet, nfeat)
}
\arguments{
\item{s_object}{Suerat Object with filtered data}

\item{normalizeMet}{The normalization method to be used}

\item{scaleF}{The Scale Factor}

\item{varianceMet}{The variance estimation method to be used}

\item{nfeat}{The number of features to be used in estimating variance}
}
\value{
Returns a Seurat object with normalized data
}
\description{
Single Cell Normalize function
}
