% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_footprintTab.R
\name{plot_tfa_per_sample}
\alias{plot_tfa_per_sample}
\title{Plot the top n TF activities per Sample in a heatmap}
\usage{
plot_tfa_per_sample(tf_activities, tf_activities_counts, tf_num)
}
\arguments{
\item{tf_activities}{a df with TF activities estimated from TF footprints}

\item{tf_activities_counts}{a df with TF per Sample}

\item{tf_num}{number of TFs with top absolute NES to be displayed}
}
\value{

}
\description{
Plot the top n TF activities per Sample in a heatmap
}
