% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_footprintTab.R
\name{plot_top_tfs}
\alias{plot_top_tfs}
\title{Plot the top n NES for the TFs}
\usage{
plot_top_tfs(tf_activities, tf_num)
}
\arguments{
\item{tf_activities}{a df with TF activities estimated from TF footprints}

\item{tf_num}{number of TFs with top absolute NES to be displayed}
}
\value{

}
\description{
Plot the top n NES for the TFs
}
