% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_compDataTab.R
\name{plotAllVenn}
\alias{plotAllVenn}
\title{Generate a Venn Diagram with DE genes}
\usage{
plotAllVenn(xlist)
}
\arguments{
\item{xlist}{A list with DE genes according to the different packages}
}
\value{
A Venn Diagram
}
\description{
Visual representation of Package Agreement/Disagreement
}
