% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_plotDataTab.R
\name{plotContrastVenn}
\alias{plotContrastVenn}
\title{Generate a Venn Diagram}
\usage{
plotContrastVenn(data, pvalue, fchange)
}
\arguments{
\item{data}{Differential Expression results (deTable)}

\item{pvalue}{P-value threshold}

\item{fchange}{Fold-Change threshold}
}
\value{
The Venn Diagram
}
\description{
Creates a Venn diagram of the DE genes between the different conditions
}
