% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_plotDataTab.R
\name{plotHeatmapTop}
\alias{plotHeatmapTop}
\title{Generate a Heatmap}
\usage{
plotHeatmapTop(data, geneNo, type, pvalue, fchange, names, session)
}
\arguments{
\item{data}{Differential Expression results (deTable)}

\item{geneNo}{The number of genes to be displayed}

\item{type}{Filter by up-, down-, or absolute significe DE}

\item{pvalue}{P-value threshold}

\item{fchange}{Fold-Change threshold}

\item{names}{Boolean - whether to show names or not}

\item{session}{Current R session}
}
\value{
Returns a heatmap
}
\description{
Generate a Heatmap
}
