% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_plotDataTab.R
\name{plotMA}
\alias{plotMA}
\title{Generate an MA plot}
\usage{
plotMA(data, expColumns, fcColumn, fcValue, pValue)
}
\arguments{
\item{data}{Differential Expression results (deTable)}

\item{expColumns}{normalized count columns of condition of interest}

\item{fcColumn}{The Fold-change column of the condition of interest}

\item{fcValue}{Fold-Change threshold}

\item{pValue}{P-value threshold}
}
\value{
The MA plot
}
\description{
Generate an MA plot
}
