% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_plotDataTab.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot PCA}
\usage{
plotPCA(data, expressionColumns)
}
\arguments{
\item{data}{Differential Expression results (deTable)}

\item{expressionColumns}{The columns with the Normalized Counts}
}
\value{
the PCA plot
}
\description{
Uses factoextra package to plot a PCA plot
}
