% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_filterDataTab.R
\name{plotScree}
\alias{plotScree}
\title{Plot Scree}
\usage{
plotScree(data)
}
\arguments{
\item{data}{Differential Expression results (deTable)}
}
\value{
Returns a Scree plot
}
\description{
Uses factoextra package to plot a Scree plot
}
