% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_filterDataTab.R
\name{qcHist}
\alias{qcHist}
\title{QC Histogram Function}
\usage{
qcHist(data)
}
\arguments{
\item{data}{Filtered/Unfiltered Count Table}
}
\value{
returns a QC Histogram
}
\description{
QC Histogram Function
}
