% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_compDataTab.R
\name{rankConsesus}
\alias{rankConsesus}
\title{Generates a ranking consesus from the selected methods}
\usage{
rankConsesus(x1_data, x2_data, x3_data, pipeline)
}
\arguments{
\item{x1_data}{DE method 1 results}

\item{x2_data}{DE method 2 results}

\item{pipeline}{The pipeline used - scRNA-Seq or Bulk}

\item{x2_data}{DE method 3 results}
}
\value{
Returns a dataframe with the ranking consensus
}
\description{
Generates a ranking consesus from the selected methods
}
