% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_goDataTab.R
\name{runGOSEQ}
\alias{runGOSEQ}
\title{Functional Annotation Pipeline (Bulk)}
\usage{
runGOSEQ(deGenes, data, genome, symbol, testType, pCorrect, pValue, appNum,
  session)
}
\arguments{
\item{deGenes}{a vector with DE gene names}

\item{data}{Filters by abs. significant, upregulation or downregulation}

\item{genome}{The genome to be used}

\item{symbol}{Type of Gene symbol}

\item{testType}{Type of Test (Cellular component, Molecular function, etc.)}

\item{pCorrect}{Whether to filter by FDR p-value or uncorrected p-value}

\item{pValue}{P-value threshold}

\item{appNum}{Format entered data appropriately to the app launched}
}
\value{
GO.wall A dataframe with Functional Annotation Results
}
\description{
Runs the GO pipeline according to the users preferences
}
