% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustTab.R
\name{sc3Cluster}
\alias{sc3Cluster}
\title{SC3 Clustering Pipeline}
\usage{
sc3Cluster(s_object, minDrop, maxDrop, estK, clustNo, nStart, session)
}
\arguments{
\item{s_object}{Seurat object with scaled and PCA data}

\item{minDrop}{Genes with percent of dropouts smaller than minDrop are filtered out before clustering.}

\item{maxDrop}{Genes with percent of dropouts larger than maxDrop are filtered out before clustering.}

\item{estK}{Boolean - whether to estimate cluster number or not}

\item{clustNo}{If estK is false - Give desired cluster number}

\item{nStart}{Minimum cells per cluster}

\item{Current}{R session}
}
\value{
Returns a list containing a Seurat object with SC3-produced clustering data and tSNE plot
}
\description{
SC3 Clustering Pipeline
}
