% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustTab.R
\name{sc_clust}
\alias{sc_clust}
\title{Single Cell Cluster Tab Server}
\usage{
sc_clust(input, output, session, normData)
}
\arguments{
\item{normData}{Reactive value containing seurat object with normalized data}
}
\value{
Returns a Reactive value containing seurat object with scaled counts and reduced dimensions (PCA data)
}
\description{
Single Cell Cluster Tab Server
}
