% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_compTab.R
\name{sc_comp}
\alias{sc_comp}
\title{SC Compare Data Server}
\usage{
sc_comp(input, output, session, finData)
}
\arguments{
\item{finData}{Clustered data results}
}
\value{
None
}
\description{
SC Compare Data Server
}
