% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_deTab.R
\name{sc_de}
\alias{sc_de}
\title{Single Cell Differential Expession Tab Server}
\usage{
sc_de(input, output, session, finData)
}
\arguments{
\item{finData}{Reactive value containing a seurat object with clustered data}
}
\value{
Returns Diffenretial Expression data
}
\description{
Single Cell Differential Expession Tab Server
}
