% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_footprintTab.R
\name{sc_dorothea_plot}
\alias{sc_dorothea_plot}
\title{Dorothea-Viper TF Activity Plot}
\usage{
sc_dorothea_plot(summarized_viper_scores_df)
}
\arguments{
\item{summarized_viper_scores_df}{a summarized df with TF activities}
}
\value{
Returns a heatmap with TF activities
}
\description{
Dorothea-Viper TF Activity Plot
}
\keyword{internal}
