% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_footprintTab.R
\name{sc_fa_Server}
\alias{sc_fa_Server}
\title{Single cell FootPrint Tab Server}
\usage{
sc_fa_Server(input, output, session, finalData)
}
\arguments{
\item{finalData}{Seurat object with dim. red. and clustering results}
}
\value{
None
}
\description{
Single cell FootPrint Tab Server
}
