% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_compTab.R
\name{sc_getAllDE}
\alias{sc_getAllDE}
\title{SC Generate Data required to compare DE Method Results}
\usage{
sc_getAllDE(data, mPCT, fc, pValue)
}
\arguments{
\item{data}{Clustering results}

\item{mPCT}{Test Genes detected in a minimum fraction of min.pct cells}

\item{fc}{Fold-change threshold}

\item{pValue}{p-value threshold}
}
\value{
Returns a list with DE genes according to the different methods
}
\description{
SC Generate Data required to compare DE Method Results
}
