% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_goTab.R
\name{sc_go}
\alias{sc_go}
\title{SC Functional Annotation Tab Server}
\usage{
sc_go(input, output, session, de, countsT)
}
\arguments{
\item{de}{Differential Expression Results (Reactive Value)}

\item{countsT}{Unfiltered Count Table (Reactive Value)}
}
\value{
None
}
\description{
SC Functional Annotation Tab Server
}
