% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_loadDataTab.R
\name{sc_loadData}
\alias{sc_loadData}
\title{Single Cell Load Data Server}
\usage{
sc_loadData(input, output, session)
}
\value{
counts - The count table to be used in the sc analysis
}
\description{
Enables the upload of two types of data: 10x Genomics and Count data
}
