% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_normTab.R
\name{sc_norm}
\alias{sc_norm}
\title{Single Cell Normalize Tab Server}
\usage{
sc_norm(input, output, session, filtData)
}
\arguments{
\item{filtData}{Reactive value containing the suerat Object with filtered data}
}
\value{
Reactive value containing Seurat object with normalized data
}
\description{
Single Cell Normalize Tab Server
}
