% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_footprintTab.R
\name{sc_progeny}
\alias{sc_progeny}
\title{Progeny Pathway Activity estimation}
\usage{
sc_progeny(s_object, organism, top, perm)
}
\arguments{
\item{s_object}{Seurat object with clustering results}

\item{organism}{Human or Mouse Model Organism}

\item{top}{No. of most variable genes per pathway}

\item{perm}{No. of permutations to be performed}
}
\value{
Returns a table with Pathway activity scores ready for display
}
\description{
Progeny Pathway Activity estimation
}
\keyword{internal}
