% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_footprintTab.R
\name{sc_progeny_plot}
\alias{sc_progeny_plot}
\title{Progeny Pathway Activity Heatmap}
\usage{
sc_progeny_plot(progeny_scores_df)
}
\arguments{
\item{progeny_scores_df}{a summarized df with Pathway activities}
}
\value{
Returns a heatmap with TF activities
}
\description{
Progeny Pathway Activity Heatmap
}
\keyword{internal}
