% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_qcTab.R
\name{sc_qc}
\alias{sc_qc}
\title{Single Cell Quality Control UI}
\usage{
sc_qc(input, output, session, countsT)
}
\arguments{
\item{countsT}{countTable loaded by loadTab}
}
\value{
A reactive value contaning the filtered data
}
\description{
Single Cell Quality Control UI
}
