% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_goTab.R
\name{sc_runGOSEQ}
\alias{sc_runGOSEQ}
\title{Functional Annotation Pipeline (Single Cell)}
\usage{
sc_runGOSEQ(deGenes, data, genome, symbol, testType, pCorrect, pValue)
}
\arguments{
\item{deGenes}{a vector with DE gene names}

\item{data}{Filters by abs. significant, upregulation or downregulation}

\item{genome}{The genome to be used}

\item{symbol}{Type of Gene symbol}

\item{testType}{Type of Test (Cellular component, Molecular function, etc.)}

\item{pCorrect}{Whether to filter by FDR p-value or uncorrected p-value}

\item{pValue}{P-value threshold}
}
\value{
GO.wall A dataframe with Functional Annotation Results
}
\description{
Runs the GO pipeline according to the users preferences
}
\details{
Only difference is in handling the first row (change it!)
}
