% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustTab.R
\name{seuratElbow}
\alias{seuratElbow}
\title{Single Cell Scale and Dimension Reduction function}
\usage{
seuratElbow(s_object)
}
\arguments{
\item{s_object}{Seurat object with normalized data}
}
\value{
Seurat object with scaled counts and reduced dimensions (PCA data)
}
\description{
Single Cell Scale and Dimension Reduction function
}
