% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustTab.R
\name{sueratClust}
\alias{sueratClust}
\title{Seurat Clustering Pipeline}
\usage{
sueratClust(s_object, dimNo, resQuant, algorithmNo, session)
}
\arguments{
\item{s_object}{Seurat object with scaled and PCA data}

\item{dimNo}{Number of dimensions to be used when clustering}

\item{resQuant}{Resolution parameter used to control the number of clusters}

\item{algorithmNo}{The clustering algorithm to be used}

\item{session}{Current R session}
}
\value{
Returns a list containing a Seurat object with clustering data and a tSNE plot
}
\description{
Seurat Clustering Pipeline
}
