% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{ui}
\alias{ui}
\title{Initiate GUI}
\format{An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 2.}
\usage{
ui
}
\value{
None
}
\description{
Initiate GUI
}
\keyword{datasets}
