% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_normTab.R
\name{variancePlotSeurat}
\alias{variancePlotSeurat}
\title{Single Cell Plot Variance Function}
\usage{
variancePlotSeurat(s_object)
}
\arguments{
\item{s_object}{Suerat Object with filtered data}
}
\value{
Variance estimation plot with the ten most variable genes
}
\description{
Single Cell Plot Variance Function
}
