% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_filterDataTab.R
\name{batchComBat}
\alias{batchComBat}
\title{ComBat Batch Effect Correction}
\usage{
batchComBat(counts, meta, session)
}
\arguments{
\item{counts, }{Filtered Count Table}

\item{meta, }{Metadata Table}

\item{counts, }{Filtered Count Table}

\item{meta, }{Metadata Table}
}
\value{
Returns batch-effect corrected table
ComBat Batch Effect Correction

Returns batch-effect corrected table
}
\description{
ComBat Batch Effect Correction
}
