% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_deDataTab.R
\name{bulk_deData}
\alias{bulk_deData}
\title{Differential Expression Tab Server}
\usage{
bulk_deData(input, output, session, fCounts)
}
\arguments{
\item{fCounts}{The reactive value containing the filtered counts}
}
\value{
A reactive value with differential expression results
}
\description{
Differential Expression Tab Server
}
