% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustTab.R
\name{clusterMonocle}
\alias{clusterMonocle}
\title{Monocle Clustering Pipeline}
\usage{
clusterMonocle(s_object, lowerDetection, redMethod, clustMethod,
  dimensionNo, estimateClust, clustNo, session)
}
\arguments{
\item{s_object}{Seurat object with scaled and PCA data}

\item{lowerDetection}{The minimum expression level that consistitutes true expression}

\item{redMethod}{Dimension reduction method to be used}

\item{dimensionNo}{Number of Dimensions to be used in clustering (tSNE)}

\item{estimateClust}{Boolean - Whether to estimate cluster No or not}

\item{clustNo}{If estimateClust is False - provide number of desired clusters}
}
\value{
Returns a list containing a Seurat object with
Monocle-produced clustering data and tSNE plot
}
\description{
Monocle Clustering Pipeline
}
