% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_deDataTab.R
\name{deEdgeR}
\alias{deEdgeR}
\title{EdgeR Differential Expression Pipeline}
\usage{
deEdgeR(readCounts, meta, testType, normMethod, useBatch)
}
\arguments{
\item{readCounts}{The filtered CountTable}

\item{meta}{Metadata table}

\item{testType}{The type of test to be used}

\item{useBatch}{If FALSE reasigns IDs to row.names}

\item{fitType}{The test method to be used}
}
\value{
EdgeR DE pipeline results
}
\description{
EdgeR Differential Expression Pipeline
}
