% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_deDataTab.R
\name{deEdgeR}
\alias{deEdgeR}
\title{EdgeR Differential Expression Pipeline}
\usage{
deEdgeR(readCounts, conditionNo, replicateNo)
}
\arguments{
\item{readCounts}{The filtered CountTable}

\item{conditionNo}{The number of Conditions}

\item{replicateNo}{The number of Replicates}
}
\value{
The results of from the differential expression table
}
\description{
EdgeR Differential Expression Pipeline
}
