% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_deDataTab.R
\name{dePipelineCaller}
\alias{dePipelineCaller}
\title{Differential Expression Pipeline Caller}
\usage{
dePipelineCaller(filteredCounts, conditionNo, replicateNo, selectedPackage,
  session)
}
\arguments{
\item{conditionNo}{The number of Conditions}

\item{replicateNo}{The number of Replicates}

\item{selectedPackage}{The package(pipeline) to be executed}

\item{session}{Current R session}

\item{readCounts}{The filtered CountTable}
}
\value{
The results of the appropriate DE pipeline
}
\description{
Function that calls the appropriate DGE pipeline
and provides exception handling
}
