% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_deDataTab.R
\name{deSequence}
\alias{deSequence}
\title{DESeq2 Differential Expression Pipeline}
\usage{
deSequence(readCounts, conditionNo, replicateNo)
}
\arguments{
\item{readCounts}{The filtered CountTable}

\item{conditionNo}{The number of Conditions}

\item{replicateNo}{The number of Replicates}
}
\value{
Returns DESeq2 DE pipeline results
}
\description{
DESeq2 Differential Expression Pipeline
}
