% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_footprintTab.R
\name{fetch_tfs_per_sample}
\alias{fetch_tfs_per_sample}
\title{TF activity per sample Sample DoRothEA}
\usage{
fetch_tfs_per_sample(de_data, reg_size)
}
\arguments{
\item{de_data}{Table with de results and symbols for gene type}

\item{reg_size}{minimun size of regulons}
}
\value{
Returns a data frame with TF activities per Sample
}
\description{
TF activity per sample Sample DoRothEA
}
