% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_filterDataTab.R
\name{filterFunction}
\alias{filterFunction}
\title{Filter Function}
\usage{
filterFunction(data, method, value)
}
\arguments{
\item{method}{Filter Option (CPM, MAX, Median)}

\item{value}{Keep only genes with counts > than this value}

\item{countTable}{The loaded Count Table}
}
\value{
countTable The filtered Count Table
}
\description{
Filter Function
}
