% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_compDataTab.R
\name{getAllDE}
\alias{getAllDE}
\title{Bulk Generate Data required to compare DE Package Results}
\usage{
getAllDE(readCounts, conditionNo, replicateNo)
}
\arguments{
\item{readCounts}{Filtered Counts Table}

\item{conditionNo}{Number of Conditions}

\item{replicateNo}{Number of Replicates}
}
\value{
Returns a list with DE genes according to the different packages
}
\description{
Bulk Generate Data required to compare DE Package Results
}
