% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_plotDataTab.R
\name{plotBarChart}
\alias{plotBarChart}
\title{Generate the Barchart}
\usage{
plotBarChart(df)
}
\arguments{
\item{df}{The Table generated by the barTable Function}
}
\value{
the Barchart
}
\description{
Generate the Barchart
}
