% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_plotDataTab.R
\name{plotMA}
\alias{plotMA}
\title{Generate an MA plot}
\usage{
plotMA(deres, fcValue, pValue)
}
\arguments{
\item{deres}{Differential Expression results (deTable)}

\item{fcValue}{Fold-Change threshold}

\item{pValue}{P-value threshold}
}
\value{
Returns a MA plot
}
\description{
Generate an MA plot
}
