% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_filterDataTab.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot PCA}
\usage{
plotPCA(data, rowNames, meta, col)
}
\arguments{
\item{data}{Differential Expression results (deTable)}

\item{rowNames, }{Boolean checking whether genes are row names or not}

\item{meta, }{metadata table}

\item{col, }{colour by batch or treatment}
}
\value{
Returns a PCA plot
}
\description{
Uses factoextra package to plot a PCA plot
}
