% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_plotDataTab.R
\name{plotScree}
\alias{plotScree}
\title{Plot Scree}
\usage{
plotScree(data, expressionColumns)
}
\arguments{
\item{data}{Differential Expression results (deTable)}

\item{expressionColumns}{The columns with the Normalized Counts}
}
\value{
Returns a Scree plot
}
\description{
Uses factoextra package to plot a Scree plot
}
