% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_plotDataTab.R
\name{plotVP}
\alias{plotVP}
\title{Generate a Volcano plot}
\usage{
plotVP(data, fcColumn, fcValue, pValue)
}
\arguments{
\item{data}{Differential Expression results (deTable)}

\item{fcColumn}{The Fold-change column of the condition of interest}

\item{fcValue}{Fold-Change threshold}

\item{pValue}{P-value threshold}
}
\value{
Retusn a Volcano plot
}
\description{
Generate a Volcano plot
}
