% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_plotDataTab.R
\name{plotVP}
\alias{plotVP}
\title{Generate a Volcano plot}
\usage{
plotVP(deres, fcValue, pValue)
}
\arguments{
\item{deres}{Differential Expression results (deTable)}

\item{fcValue}{Fold-Change threshold}

\item{pValue}{P-value threshold}
}
\value{
Returns a Volcano plot
}
\description{
Generate a Volcano plot
}
