% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustTab.R
\name{sc3Cluster}
\alias{sc3Cluster}
\title{SC3 Clustering Pipeline}
\usage{
sc3Cluster(s_object, estK, clustNo, cellsPerC)
}
\arguments{
\item{s_object}{Seurat object with scaled and PCA data}

\item{estK}{Boolean - whether to estimate cluster number or not}

\item{clustNo}{If estK is false - Give desired cluster number}

\item{cellsPerC}{Minimum cells per cluster}
}
\value{
Seurat object with SC3 clustering data
}
\description{
SC3 Clustering Pipeline
}
