% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_footprintTab.R
\name{sc_fa_Server}
\alias{sc_fa_Server}
\title{Bulk FootPrint Tab Server}
\usage{
sc_fa_Server(input, output, session, counts, de)
}
\arguments{
\item{counts}{Unfiltered Count Table (Reactive Value)}

\item{de}{Differential Expression Results (Reactive Value)}
}
\value{
None
}
\description{
Bulk FootPrint Tab Server
}
