% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_qcTab.R
\name{seuratQC}
\alias{seuratQC}
\title{Single Cell Quality Control}
\usage{
seuratQC(seurat_object, minF, maxF, session)
}
\arguments{
\item{seurat_object}{The initialized seurat object with the count data}

\item{minF}{Minimum feature counts per cell}

\item{maxF}{Maximum feature counts per cell}

\item{session}{Current R session supplied by the server}
}
\value{
A Seurat_object with the filtered data
}
\description{
Performs QC on Seurat object, which serves as a backbone for the sc pipeline
}
