% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustTab.R
\name{sueratClust}
\alias{sueratClust}
\title{Seurat Clustering Pipeline}
\usage{
sueratClust(s_object, dimNo, resQuant, algorithmNo)
}
\arguments{
\item{s_object}{Seurat object with scaled and PCA data}
}
\value{
Seurat object with clustering data
}
\description{
Seurat Clustering Pipeline
}
